refactor: (#77) 레포지토리 테스트 어노테이션 리팩토링
