+++ backend/src/test/java/com/votogether/domain/category/repository/CategoryRepositoryTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import com.votogether.domain.RepositoryTest;
+import com.votogether.RepositoryTest;
 import com.votogether.domain.category.entity.Category;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
